added activity summary event
