+++ lib/pangea/activity_planner/activity_status_message.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+
+import 'package:matrix/matrix.dart';
+
+import 'package:fluffychat/config/themes.dart';
+import 'package:fluffychat/pangea/activity_planner/activity_finished_status_message.dart';
+import 'package:fluffychat/pangea/activity_planner/activity_room_extension.dart';
+import 'package:fluffychat/pangea/activity_planner/activity_unfinished_status_message.dart';
+
+class ActivityStatusMessage extends StatelessWidget {
+  final Room room;
+
+  const ActivityStatusMessage({
+    super.key,
+    required this.room,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Positioned(
+      left: 0,
+      right: 0,
+      bottom: 0,
+      child: Material(
+        child: AnimatedSize(
+          duration: FluffyThemes.animationDuration,
+          child: !room.hasJoinedActivity || room.activityIsFinished
+              ? Padding(
+                  padding: EdgeInsets.only(
+                    bottom: FluffyThemes.isColumnMode(context) ? 32.0 : 16.0,
+                    left: 16.0,
+                    right: 16.0,
+                  ),
+                  child: ConstrainedBox(
+                    constraints: BoxConstraints(
+                      maxHeight: MediaQuery.of(context).size.height * 0.8,
+                    ),
+                    child: SingleChildScrollView(
+                      child: room.activityIsFinished
+                          ? ActivityFinishedStatusMessage(room: room)
+                          : ActivityUnfinishedStatusMessage(room: room),
+                    ),
+                  ),
+                )
+              : const SizedBox.shrink(),
+        ),
+      ),
+    );
+  }
+}
